requirements.txt
fastapi>=0.111
uvicorn[standard]>=0.29
SQLAlchemy>=2.0
python-dotenv>=1.0.1
pydantic[email]>=2.7
psycopg[binary]>=3.2
openai>=1.40
python-multipart
pdfplumber
python-docx
python-pptx
google-generativeai